requirements.txt
fastapi
uvicorn[standard]
pydantic
pydantic-settings
python-dotenv
google-generativeai
logfire[fastapi,httpx]
httpx
python-multipart
Pillow
supabase